+++ src/main/kotlin/happybeans/service/RestaurantReviewService.kt
@@ -9,12 +9,14 @@ import happybeans.model.User
 import happybeans.repository.RestaurantRepository
 import happybeans.repository.RestaurantReviewRepository
 import jakarta.persistence.EntityNotFoundException
+import jakarta.transaction.Transactional
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import java.math.RoundingMode
 
+@Transactional
 @Service
 class RestaurantReviewService(
     private val restaurantReviewRepository: RestaurantReviewRepository,
